vision_bot/requirements.txt
discord.py>=2.3.2,<3.0.0
python-dotenv>=1.0.1,<2.0.0
aiohttp>=3.9.0
